using prettier for code formatting

+++ contracts/Migrations.sol
@@ -1,10 +1,9 @@
 pragma solidity ^0.5.8;
 
-
 //actually this contract is only needed that "npm run coverage" is working
 contract Migrations {
     address public owner;
-    uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
     constructor() public {
         owner = msg.sender;
@@ -14,7 +13,7 @@ contract Migrations {
         if (msg.sender == owner) _;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         last_completed_migration = completed;
     }
